+++ app/routes/components/Popover/demos/Main/index.tsx
@@ -5,7 +5,7 @@ import { Paper } from "@/components/Paper";
 import { type Placement, Popover } from "@/components/Popover";
 import { Select } from "@/components/Select";
 
-import styles from "./Demo01.module.scss";
+import styles from "./index.module.css";
 
 export default function Demo() {
   const [placement, setPlacement] = useState(null as Placement | null);