+++ src/components/Popover/index.ts
@@ -1,4 +1,4 @@
-// Popover from copy-ui @ 2025-11-28.
+// Popover from copy-ui @ 2025-12-14.
 
 export { type Placement, Popover } from "./popover";
 export type { PopoverTriggerRender } from "./popover-trigger";